fix: change background color
